requirements.txt
pandas>=2.0.0,<3.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0,<1.0.0
pydantic>=2.0.0,<3.0.0
joblib>=1.3.0,<2.0.0
pytest>=7.4.0,<9.0.0
httpx>=0.24.0,<1.0.0
PyYAML>=6.0,<7.0
